restructure survey insights into multiple files
